Cubit folder reorganization
