+++ lib/pages/caregiver/plans_page.dart
@@ -1,7 +1,7 @@
 import 'dart:developer';
 
 import 'package:flutter/material.dart';
-import 'package:fokus/logic/caregiver_plans/caregiver_plans_cubit.dart';
+import 'package:fokus/logic/caregiver_plans_cubit.dart';
 import 'package:fokus/model/ui/app_page.dart';
 import 'package:fokus/model/ui/plan/ui_plan.dart';
 import 'package:fokus/services/app_locales.dart';
