+++ lib/pages/child/panel_page.dart
@@ -3,7 +3,7 @@ import 'dart:developer';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'package:fokus/logic/child_plans/child_plans_cubit.dart';
+import 'package:fokus/logic/child_plans_cubit.dart';
 import 'package:fokus/model/ui/app_page.dart';
 import 'package:fokus/logic/timer/timer_cubit.dart';
 import 'package:fokus/model/db/plan/plan_instance_state.dart';
@@ -36,7 +36,7 @@ class _ChildPanelPageState extends State<ChildPanelPage> {
 	      children: [
 	        ChildCustomHeader(),
 		      LoadableBlocBuilder<ChildPlansCubit>(
-				      builder: (context, state) => AppSegments(segments: _buildPanelSegments(state))
+				    builder: (context, state) => AppSegments(segments: _buildPanelSegments(state))
 		      ),
 	        Row(
 						mainAxisAlignment: MainAxisAlignment.end,
